UserService.cs added
