+++ PizzaApp/Controllers/UsersController.cs
@@ -9,17 +9,17 @@ namespace PizzaApp.Controllers;
 
 public class UsersController : BaseApiController
 {
-    private readonly IUnitOfWork _unitOfWork;
+    private readonly IUserService _userService;
 
-    public UsersController(IUnitOfWork unitOfWork )
+    public UsersController(IUserService userService )
     {
-        _unitOfWork = unitOfWork;
+        _userService = userService;
     }
 
     [HttpGet]
     [Authorize]
     public async Task<IEnumerable<User>> GetUsers()
     {
-        return await _unitOfWork.UserRepository.GetUsers();
+        return await _userService.GetUsers();
     }
 }
