+++ PizzaApp/Interfaces/IUserRepository.cs
@@ -4,6 +4,6 @@ namespace PizzaApp.Interfaces;
 
 public interface IUserRepository
 {
-    Task<IEnumerable<User>> GetUsers();
+    Task<IQueryable<User>> GetUsers();
     Task<User> GetUserByName(string name);
 }
