+++ PizzaApp/Interfaces/IUserService.cs
@@ -1,0 +1,10 @@
+﻿using PizzaApp.DTOs;
+using PizzaApp.Entities;
+
+namespace PizzaApp.Interfaces;
+
+public interface IUserService
+{
+    Task<IEnumerable<User>> GetUsers();
+
+}
