+++ PizzaApp/Services/UserService.cs
@@ -1,0 +1,23 @@
+﻿using AutoMapper;
+using PizzaApp.DTOs;
+using PizzaApp.Entities;
+using PizzaApp.Interfaces;
+
+namespace PizzaApp.Services;
+
+public class UserService : IUserService
+{
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly IMapper _mapper;
+
+    public UserService(IUnitOfWork unitOfWork, IMapper mapper)
+    {
+        _unitOfWork = unitOfWork;
+        _mapper = mapper;
+    }
+
+    public async Task<IEnumerable<User>> GetUsers()
+    {
+        return await _unitOfWork.UserRepository.GetUsers();
+    }
+}
